# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema TKD API
  version: 1.0.0
  description: Contrato del sistema. PostgreSQL es la fuente de verdad y la escuela se obtiene exclusivamente de la sesión.
servers:
  - url: /api
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tkd_session
  parameters:
    Csrf:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 20 }
    Idempotency:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Conflict:
      description: La versión cambió o la operación contradice el estado actual.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
    ResourceCreated:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
        replayed: { type: boolean, default: false }
...
    StudentWrite:
      type: object
      additionalProperties: false
      required: [firstNames,lastNames,birthDate,gradeId,primaryName,primaryRelation,primaryPhone,emergencyName,emergencyRelation,emergencyPhone,enrollmentDate,billingDay]
      properties:
        firstNames: { type: string, minLength: 1 }
        lastNames: { type: string, minLength: 1, description: Admite un solo apellido. }
        birthDate: { type: string, format: date }
        gradeId: { type: string, format: uuid }
        email: { type: string, format: email }
        emailOwner: { type: string, enum: [alumno, responsable, ''] }
        primaryName: { type: string, minLength: 1 }
        primaryRelation: { type: string, minLength: 1 }
        primaryPhone: { type: string, minLength: 1 }
        emergencyName: { type: string, minLength: 1 }
        emergencyRelation: { type: string, minLength: 1 }
        emergencyPhone: { type: string, minLength: 1 }
        restrictions: { type: string }
        healthConsent: { type: boolean, description: Consentimiento registrado en este entorno. }
        enrollmentDate: { type: string, format: date }
        billingDay: { type: integer, minimum: 1, maximum: 31 }
        version: { type: integer, minimum: 1 }
...
    NoteWrite:
      type: object
      additionalProperties: false
      required: [date,topic,text]
      properties:
        date: { type: string, format: date }
        topic: { type: string, minLength: 1 }
        text: { type: string, minLength: 1 }
        sessionId: { type: [string,'null'], format: uuid }
        version: { type: integer, minimum: 1 }
    GroupWrite:
      type: object
      additionalProperties: false
      required: [name,days,start,end]
      properties:
        name: { type: string, minLength: 1 }
        days: { type: array, minItems: 1, uniqueItems: true, items: { type: string, enum: [Lunes,Martes,Miércoles,Jueves,Viernes,Sábado] } }
        start: { type: string, format: time }
        end: { type: string, format: time }
        version: { type: integer, minimum: 1 }
    AssignmentWrite:
      type: object
      additionalProperties: false
      required: [studentId,assignedFrom]
      properties:
        studentId: { type: string, format: uuid }
        assignedFrom: { type: string, format: date }
    ParticipantWrite:
      type: object
      additionalProperties: false
      required: [studentId]
      properties:
        studentId: { type: string, format: uuid }
    SessionWrite:
      type: object
      additionalProperties: false
      required: [groupId,date,start,end]
      properties:
        groupId: { type: string, format: uuid }
        date: { type: string, format: date }
        start: { type: string, format: time }
        end: { type: string, format: time }
...
paths:
...
  /students:
    post:
      operationId: createStudent
      parameters: [{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudentWrite' } } } }
      responses: { '201': { description: Alumno creado, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceCreated' } } } }, '200': { description: Reintento idempotente del alta }, '400': { description: Validación incorrecta } }
  /students/{id}:
    put: { operationId: updateStudent, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudentWrite' } } } }, responses: { '204': { description: Actualizado }, '409': { $ref: '#/components/responses/Conflict' } } }
...
  /students/{id}/notes:
    post: { operationId: createNote, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NoteWrite' } } } }, responses: { '201': { description: Nota creada }, '200': { description: Reintento idempotente del alta } } }
  /notes/{id}:
    put: { operationId: updateNote, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NoteWrite' } } } }, responses: { '204': { description: Nota actualizada }, '409': { $ref: '#/components/responses/Conflict' } } }
    delete: { operationId: deleteNote, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ name: version, in: query, required: true, schema: { type: integer, minimum: 1 } },{ $ref: '#/components/parameters/Csrf' }], responses: { '204': { description: Nota eliminada }, '409': { $ref: '#/components/responses/Conflict' } } }
  /groups:
    post: { operationId: createGroup, parameters: [{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GroupWrite' } } } }, responses: { '201': { description: Grupo creado }, '200': { description: Reintento idempotente del alta } } }
  /groups/{id}:
    put: { operationId: updateGroup, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GroupWrite' } } } }, responses: { '204': { description: Grupo actualizado } } }
  /groups/{id}/assignments:
    post: { operationId: assignStudent, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentWrite' } } } }, responses: { '204': { description: Asignación vigente } } }
  /sessions:
    post: { operationId: createSession, parameters: [{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SessionWrite' } } } }, responses: { '201': { description: Sesión creada o recuperada } } }
  /sessions/{id}/participants:
    post: { operationId: inviteParticipant, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ParticipantWrite' } } } }, responses: { '204': { description: Participante presente sin duplicarse } } }
...